Remove unhighlighted comments by default
This implements most of the feature request in https://github.com/experimatt/hn-jobs-focus/issues/1 , but instead of adding it as an option, it removes unhighlighted comments by default.

Feel free to add it in the UI as a toggle or whatever you see fit.  Thanks for the cool little time-saver of an extension!